mcp: carve state.rs (pure move)
McpAppState and its satellites (SessionStatusState,
InsufficientRewindNotice, DensityMaintenanceSatisfaction,
PendingApprovalState), the TaskLauncher seam, SharedMcpState, and the
phase/verbosity string helpers: 1,320 body lines and 9 tests verbatim.

Field widening beyond the usual item widening: fields of the moved
structs that sibling mcp modules (and their tests) access directly —
SessionStatusState's four, all of InsufficientRewindNotice's (sibling
tests construct it literally), and McpAppState's overrides,
session_aliases, session_usage, rewind bookkeeping, and
user_display_activation_pending.

Purity-checked; mcp:: tests 178/178.
